server/requirements.txt
Flask
Flask-SocketIO
Flask-Migrate
Flask-SQLAlchemy
pyngrok
redis
python-dotenv
Werkzeug
PyJWT